+++ neuwon/__init__.py
@@ -3,3 +3,74 @@ NEUWON is a simulation framework for neuroscience and artificial intelligence
 specializing in conductance based models. This software is a modern remake of
 the NEURON simulator. It is accurate, efficient, and easy to use.
 """
+
+class Species:
+    """ """
+    def __init__(self, name,
+            charge = 0,
+            transmembrane = False,
+            reversal_potential = "nerst",
+            intra_concentration = 0.0,
+            extra_concentration = 0.0,
+            intra_diffusivity = None,
+            extra_diffusivity = None,
+            intra_decay_period = float("inf"),
+            extra_decay_period = float("inf")):
+        """
+        If diffusivity is not given, then the concentration is constant.
+        Argument reversal_potential is one of: number, "nerst", "goldman_hodgkin_katz"
+        """
+        self.name = str(name)
+        self.charge = int(charge)
+        self.transmembrane = bool(transmembrane)
+        self.intra_concentration = float(intra_concentration)
+        self.extra_concentration = float(extra_concentration)
+        self.intra_diffusivity = float(intra_diffusivity) if intra_diffusivity is not None else None
+        self.extra_diffusivity = float(extra_diffusivity) if extra_diffusivity is not None else None
+        self.intra_decay_period = float(intra_decay_period)
+        self.extra_decay_period = float(extra_decay_period)
+        assert(self.intra_concentration >= 0.0)
+        assert(self.extra_concentration >= 0.0)
+        assert(self.intra_diffusivity is None or self.intra_diffusivity >= 0)
+        assert(self.extra_diffusivity is None or self.extra_diffusivity >= 0)
+        assert(self.intra_decay_period > 0.0)
+        assert(self.extra_decay_period > 0.0)
+        if reversal_potential == "nerst":
+            self.reversal_potential = str(reversal_potential)
+            self._reversal_potential_method = lambda T, i, o, v: nerst_potential(self.charge, T, i, o)
+        elif reversal_potential == "goldman_hodgkin_katz":
+            self.reversal_potential = str(reversal_potential)
+            self._reversal_potential_method = self.goldman_hodgkin_katz
+        else:
+            self.reversal_potential = float(reversal_potential)
+            self._reversal_potential_method = lambda T, i, o, v: self.reversal_potential
+
+class Reaction:
+    """ Abstract class for specifying reactions and mechanisms. """
+    @classmethod
+    def name(self):
+        """ A unique name for this reaction and all of its instances. """
+        raise TypeError("Abstract method called by %s."%repr(self))
+
+    @classmethod
+    def initialize(self, database):
+        """ (Optional) This method is called after the Model has been created.
+        This method is called on a deep copy of each Reaction object.
+
+        Argument database is a function(name) -> value
+
+        (Optional) Returns a new Reaction object to use in place of this one. """
+        pass
+
+    @classmethod
+    def new_instances(self, database_access, locations, *args, **kw_args):
+        """ """
+        pass
+
+    @classmethod
+    def advance(self, database_access):
+        """ Advance all instances of this reaction.
+
+        Argument database_access is function: f(component_name) -> value
+        """
+        raise TypeError("Abstract method called by %s."%repr(self))
