+++ neuwon/api/__init__.py
@@ -1,15 +1,13 @@
 __all__ = """
-AccessHandle
 Location
 Model
 Neighbor
 Reaction
 Real
-Segment
 Species
 """.split()
 
-from neuwon.common import Real, Location, AccessHandle
+from neuwon.common import Real, Location
 from neuwon.model import Model
 from neuwon.reactions import Reaction
 from neuwon.segments import Segment
