+++ neuwon/species.py
@@ -47,93 +47,6 @@ library = {
     },
 }
 
-class Species:
-    """ """
-    def __init__(self, name,
-            charge = 0,
-            transmembrane = False,
-            reversal_potential = "nerst",
-            intra_concentration = 0.0,
-            extra_concentration = 0.0,
-            intra_diffusivity = None,
-            extra_diffusivity = None,
-            intra_decay_period = float("inf"),
-            extra_decay_period = float("inf")):
-        """
-        If diffusivity is not given, then the concentration is constant.
-        Argument reversal_potential is one of: number, "nerst", "goldman_hodgkin_katz"
-        """
-        self.name = str(name)
-        self.charge = int(charge)
-        self.transmembrane = bool(transmembrane)
-        self.intra_concentration = float(intra_concentration)
-        self.extra_concentration = float(extra_concentration)
-        self.intra_diffusivity = float(intra_diffusivity) if intra_diffusivity is not None else None
-        self.extra_diffusivity = float(extra_diffusivity) if extra_diffusivity is not None else None
-        self.intra_decay_period = float(intra_decay_period)
-        self.extra_decay_period = float(extra_decay_period)
-        assert(self.intra_concentration >= 0.0)
-        assert(self.extra_concentration >= 0.0)
-        assert(self.intra_diffusivity is None or self.intra_diffusivity >= 0)
-        assert(self.extra_diffusivity is None or self.extra_diffusivity >= 0)
-        assert(self.intra_decay_period > 0.0)
-        assert(self.extra_decay_period > 0.0)
-        if reversal_potential == "nerst":
-            self.reversal_potential = str(reversal_potential)
-            self._reversal_potential_method = lambda T, i, o, v: nerst_potential(self.charge, T, i, o)
-        elif reversal_potential == "goldman_hodgkin_katz":
-            self.reversal_potential = str(reversal_potential)
-            self._reversal_potential_method = self.goldman_hodgkin_katz
-        else:
-            self.reversal_potential = float(reversal_potential)
-            self._reversal_potential_method = lambda T, i, o, v: self.reversal_potential
-
-def species_advance(model):
-    """ Note: Each call to this method integrates over half a time step. """
-    dt = model._electrics.time_step
-    # Accumulate the net conductances and driving voltages from the chemical data.
-    # model._electrics.conductances     = cp.zeros(len(geometry), dtype=Real)
-    # model._electrics.driving_voltages = cp.zeros(len(geometry), dtype=Real)
-    model._electrics.conductances.fill(0)     # Zero accumulator.
-    model._electrics.driving_voltages.fill(0) # Zero accumulator.
-    T = model.celsius + 273.15
-    for s in model._species.values():
-        if not s.transmembrane: continue
-        s.reversal_potential = s._reversal_potential_method(
-            T,
-            s.intra_concentration if s.intra is None else s.intra.concentrations,
-            s.extra_concentration if s.extra is None else s.extra.concentrations,
-            model._electrics.voltages)
-        model._electrics.conductances += s.conductances
-        model._electrics.driving_voltages += s.conductances * s.reversal_potential
-    model._electrics.driving_voltages /= model._electrics.conductances
-    model._electrics.driving_voltages = cp.nan_to_num(model._electrics.driving_voltages)
-    # Calculate the transmembrane currents.
-    diff_v = model._electrics.driving_voltages - model._electrics.voltages
-    recip_rc = model._electrics.conductances / model._electrics.capacitances
-    alpha = cp.exp(-dt * recip_rc)
-    model._electrics.voltages += diff_v * (1.0 - alpha)
-    # Calculate the lateral currents throughout the neurons.
-    model._electrics.voltages = model._electrics.irm.dot(model._electrics.voltages)
-    # Calculate the transmembrane ion flows.
-    for s in model._species.values():
-        if not s.transmembrane: continue
-        if s.intra is None and s.extra is None: continue
-        integral_v = dt * (s.reversal_potential - model._electrics.driving_voltages)
-        integral_v += rc * diff_v * alpha
-        moles = s.conductances * integral_v / (s.charge * F)
-        if s.intra is not None:
-            s.intra.concentrations += moles / model.geometry.intra_volumes
-        if s.extra is not None:
-            s.extra.concentrations -= moles / model.geometry.extra_volumes
-    # Calculate the local release / removal of chemicals.
-    for s in model._species.values():
-        for x in (s.intra, s.extra):
-            if x is None: continue
-            x.concentrations = cp.maximum(0, x.concentrations + x.release_rates * dt)
-            # Calculate the lateral diffusion throughout the space.
-            x.concentrations = x.irm.dot(x.concentrations)
-
 class _Diffusion:
     def __init__(self, time_step, geometry, species, where):
         self.time_step = time_step
